move methods which are unrelated to Packit packages/env from symlinker to symlink utils
