+++ src/platforms/symlink.rs
@@ -1,16 +1,7 @@
 use std::path::Path;
 
-use thiserror::Error;
-
-/// The errors that occur during display.
-#[derive(Error, Debug)]
-pub enum SymlinkError {
-    #[error("Symlink IO failed")]
-    IOError(#[from] std::io::Error),
-
-    #[error("Path is not a symlink")]
-    NonSymlink,
-}
+#[cfg(any(target_os = "macos", target_os = "linux"))]
+use crate::utils::symlink::SymlinkError;
 
 #[cfg(any(target_os = "macos", target_os = "linux"))]
 pub fn create_symlink(source: &Path, destination: &Path) -> Result<(), SymlinkError> {
