+++ src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod constants;
 pub mod env;
+pub mod symlink;
 pub mod unwrap_or_exit;
